refactor: extract dispatch helpers from cli.py to cli_helper.py
Move _detect_file_type, _pull_file, _push_file, _pull_folder,
_push_folder out of cli.py into cli_helper.py. cli.py should focus
on UX and wiring; dispatch logic belongs elsewhere.

cli.py: 3382 → 2689 lines
cli_helper.py: 725 lines (new)
